Simplify print_status in rofi-bluetooth
This replaces the mapfile + bash array code with a simple pipe into a while loop and simplifies the logic a bit.

It also makes the script more portable. The only remaining bashism is the pair of  echo -e "$options"  calls, which can be replaced with a slightly uglier printf "$options" if posix compatibility is desired. Then the script runs fine on dash.